--- a/src/Datadog.Trace/Configuration/TracerSettings.cs
+++ b/src/Datadog.Trace/Configuration/TracerSettings.cs
@@ -22,9 +22,6 @@ namespace Datadog.Trace.Configuration
         /// </summary>
         public const int DefaultAgentPort = 8126;
 
-        private IDictionary<string, string> _headerTags;
-        private Func<bool> _headerTagsEmpty;
-
         /// <summary>
         /// Initializes a new instance of the <see cref="TracerSettings"/> class with default values.
         /// </summary>
@@ -247,23 +244,7 @@ namespace Datadog.Trace.Configuration
         /// <summary>
         /// Gets or sets the map of header keys to tag names, which are applied to the root <see cref="Span"/> of incoming requests.
         /// </summary>
-        public IDictionary<string, string> HeaderTags
-        {
-            get => _headerTags;
-            set
-            {
-                _headerTags = value;
-
-                if (value is ConcurrentDictionary<string, string> concurrentDictionary)
-                {
-                    _headerTagsEmpty = () => concurrentDictionary.IsEmpty;
-                }
-                else
-                {
-                    _headerTagsEmpty = () => value.Count == 0;
-                }
-            }
-        }
+        public IDictionary<string, string> HeaderTags { get; set; }
 
         /// <summary>
         /// Gets or sets the port where the DogStatsd server is listening for connections.
@@ -289,7 +270,7 @@ namespace Datadog.Trace.Configuration
         /// </summary>
         public bool StartupDiagnosticLogEnabled { get; set; }
 
-        internal bool HeaderTagsEmpty => _headerTagsEmpty();
+        internal bool HeaderTagsEmpty => HeaderTags is ConcurrentDictionary<string, string> concurrent ? concurrent.IsEmpty : HeaderTags.Count == 0;
 
         /// <summary>
         /// Create a <see cref="TracerSettings"/> populated from the default sources